Extract presentational dashboard components
